Configurable precision argument
The default setup previously would produce distributions that were skewed by >5%. This is due to the values not being evenly distributed within the hash ring.

This adds a second argument when creating a HashRing that specifies a multiple for how many values should be generated within the hash ring space. The prior default was `40`, the new default is `1000`. This takes a bit more memory and a bit more CPU upon HashRing creation, but produces less skewed distributions.